+++ packages/client/base/src/services/hosted/v3/crossmintHostedCheckoutV3Service.ts
@@ -1,4 +1,8 @@
-import type { CrossmintHostedCheckoutV3Props } from "@/types/hosted/v3/CrossmintHostedCheckoutV3Props";
+import {
+    type CrossmintHostedCheckoutV3AllProps,
+    isHostedCheckoutV3ExistingOrderProps,
+    type CrossmintHostedCheckoutV3OrderProps,
+} from "@/types/hosted/v3/CrossmintHostedCheckoutV3Props";
 import { appendObjectToQueryParams } from "@/utils/appendObjectToQueryParams";
 import { NewTabWindow, PopupWindow } from "@crossmint/client-sdk-window";
 import type { CrossmintApiClient } from "@crossmint/common-sdk-base";
@@ -6,7 +10,7 @@ import { crossmintHostedCheckoutOverlayService } from "./crossmintHostedCheckout
 
 export type CrossmintHostedCheckoutV3ServiceProps = {
     apiClient: CrossmintApiClient;
-    hostedCheckoutProps: CrossmintHostedCheckoutV3Props;
+    hostedCheckoutProps: CrossmintHostedCheckoutV3AllProps;
 };
 
 export function crossmintHostedCheckoutV3Service({
@@ -15,11 +19,20 @@ export function crossmintHostedCheckoutV3Service({
 }: CrossmintHostedCheckoutV3ServiceProps) {
     const overlayService = crossmintHostedCheckoutOverlayService();
 
-    function getUrl(props: CrossmintHostedCheckoutV3Props) {
-        const urlWithPath = apiClient.buildUrl("/sdk/2024-03-05/hosted-checkout");
+    function getUrl(props: CrossmintHostedCheckoutV3AllProps) {
+        const isExistingOrder = isHostedCheckoutV3ExistingOrderProps(props);
+        const path = isExistingOrder
+            ? `/sdk/2024-03-05/hosted-checkout/${encodeURIComponent(props.orderId)}`
+            : "/sdk/2024-03-05/hosted-checkout";
+        const urlWithPath = apiClient.buildUrl(path);
         const queryParams = new URLSearchParams();
 
-        appendObjectToQueryParams(queryParams, props);
+        if (isExistingOrder) {
+            const { orderId, ...restOfParams } = props as CrossmintHostedCheckoutV3OrderProps;
+            appendObjectToQueryParams(queryParams, restOfParams);
+        } else {
+            appendObjectToQueryParams(queryParams, props);
+        }
 
         queryParams.append("apiKey", apiClient.crossmint.apiKey);
         queryParams.append("sdkMetadata", JSON.stringify(apiClient["internalConfig"].sdkMetadata));